Fix control flow handling for ubergraphs
Ubergraphs sometimes push execution flow before the computed jump. As
we started analysis at the "entrypoints", we had no way to know about
this, so the pop execution flow nodes would not be connected. Instead
of that, use the "entrypoint" addr as the target of the computed jump
in order to simulate what the actual execution does.

In the process of understanding what this atrocity of a codebase does
once again, even though I wrote all of it myself, I found some things
that could be improved. So I improved them.